# openapi/openapi.yml
components:
  schemas:
...
    EndpointEntryType:
      description: 'Entry endpoint resource types, mapping to their pydantic models
        from the

        `optimade` package.'
      enum:
      - references
      - structures
      title: EndpointEntryType
...
    ErrorResponse:
      description: errors MUST be present and data MUST be skipped
      properties:
        data:
          anyOf:
          - $ref: '#/components/schemas/Resource'
          - items:
              $ref: '#/components/schemas/Resource'
            type: array
          description: Outputted Data
          title: Data
          uniqueItems: true
        errors:
          description: A list of OPTIMADE-specific JSON API error objects, where the
            field detail MUST be present.
          items:
            $ref: '#/components/schemas/OptimadeError'
          title: Errors
          type: array
          uniqueItems: true
        included:
          description: A list of unique included resources
          items:
            $ref: '#/components/schemas/Resource'
          title: Included
          type: array
          uniqueItems: true
        jsonapi:
          allOf:
          - $ref: '#/components/schemas/JsonApi'
          description: Information about the JSON API used
          title: Jsonapi
        links:
          allOf:
          - $ref: '#/components/schemas/ToplevelLinks'
          description: Links associated with the primary data or errors
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/ResponseMeta'
          description: A meta object containing non-standard information.
          title: Meta
      required:
      - meta
      - errors
      title: ErrorResponse
      type: object
...
    JsonApi:
      description: An object describing the server's implementation
      properties:
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: Non-standard meta information
          title: Meta
        version:
          default: '1.0'
          description: Version of the json API used
          title: Version
          type: string
      title: JsonApi
      type: object
...
    Meta:
      description: Non-standard meta-information that can not be represented as an
        attribute or relationship.
      properties: {}
      title: Meta
      type: object
...
    QueriesResponse:
      description: Successful response for `GET /gateways/{gateway_ID}/queries`.
      properties:
        data:
          description: List of unique OPTIMADE gateway query resource objects.
          items:
            $ref: '#/components/schemas/QueryResource'
          title: Data
          type: array
          uniqueItems: true
        errors:
          description: A list of unique errors
          items:
            $ref: '#/components/schemas/Error'
          title: Errors
          type: array
          uniqueItems: true
        included:
          anyOf:
          - items:
              $ref: '#/components/schemas/optimade__models__entries__EntryResource'
            type: array
          - items:
              type: object
            type: array
          title: Included
          uniqueItems: true
        jsonapi:
          allOf:
          - $ref: '#/components/schemas/JsonApi'
          description: Information about the JSON API used
          title: Jsonapi
        links:
          allOf:
          - $ref: '#/components/schemas/ToplevelLinks'
          description: Links associated with the primary data or errors
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/ResponseMeta'
          description: A meta object containing non-standard information
          title: Meta
      required:
      - data
      - meta
      title: QueriesResponse
      type: object
    QueriesResponseSingle:
      description: 'Successful response for `POST /gateways/{gateway_ID}/queries`

        and `GET /gateways/{gateway_ID}/queries/{query_id}`.'
      properties:
        data:
          allOf:
          - $ref: '#/components/schemas/QueryResource'
          description: 'A unique OPTIMADE gateway query resource object.

            The OPTIMADE gateway query resource object has just been created or found
            according to the specific query parameter(s) or URL id.'
          title: Data
        errors:
          description: A list of unique errors
          items:
            $ref: '#/components/schemas/Error'
          title: Errors
          type: array
          uniqueItems: true
        included:
          anyOf:
          - items:
              $ref: '#/components/schemas/optimade__models__entries__EntryResource'
            type: array
          - items:
              type: object
            type: array
          title: Included
          uniqueItems: true
        jsonapi:
          allOf:
          - $ref: '#/components/schemas/JsonApi'
          description: Information about the JSON API used
          title: Jsonapi
        links:
          allOf:
          - $ref: '#/components/schemas/ToplevelLinks'
          description: Links associated with the primary data or errors
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/ResponseMeta'
          description: A meta object containing non-standard information
          title: Meta
      required:
      - data
      - meta
      title: QueriesResponseSingle
      type: object
    QueryCreate:
      description: Model for creating new Query resources in the MongoDB
      properties:
        endpoint:
          $ref: '#/components/schemas/EndpointEntryType'
        gateway_id:
          description: The OPTIMADE gateway ID for this query.
          title: Gateway Id
          type: string
        id:
          title: Id
          type: string
        immutable_id:
          description: "The entry's immutable ID (e.g., an UUID). This is important\
            \ for databases having preferred IDs that point to \"the latest version\"\
            \ of a record, but still offer access to older variants. This ID maps\
            \ to the version-specific record, in case it changes in the future.\n\n\
            - **Type**: string.\n\n- **Requirements/Conventions**:\n    - **Support**:\
            \ OPTIONAL support in implementations, i.e., MAY be `null`.\n    - **Query**:\
            \ MUST be a queryable property with support for all mandatory filter features.\n\
            \n- **Examples**:\n    - `\"8bd3e750-b477-41a0-9b11-3a799f21b44f\"`\n\
            \    - `\"fjeiwoj,54;@=%<>#32\"` (Strings that are not URL-safe are allowed.)"
          title: Immutable Id
          type: string
        last_modified:
          format: date-time
          title: Last Modified
          type: string
        query_parameters:
          allOf:
          - $ref: '#/components/schemas/OptimadeQueryParameters'
          description: OPTIMADE query parameters for entry listing endpoints used
            for this query.
          title: Query Parameters
          type: object
        response:
          allOf:
          - $ref: '#/components/schemas/GatewayQueryResponse'
          description: Response from gateway query.
          title: Response
        state:
          $ref: '#/components/schemas/QueryState'
      required:
      - gateway_id
      - query_parameters
      title: QueryCreate
      type: object
    QueryResource:
      description: OPTIMADE query resource for a gateway
      properties:
        attributes:
          $ref: '#/components/schemas/QueryResourceAttributes'
        id:
          description: "An entry's ID as defined in section Definition of Terms.\n\
            \n- **Type**: string.\n\n- **Requirements/Conventions**:\n    - **Support**:\
            \ MUST be supported by all implementations, MUST NOT be `null`.\n    -\
            \ **Query**: MUST be a queryable property with support for all mandatory\
            \ filter features.\n    - **Response**: REQUIRED in the response.\n\n\
            - **Examples**:\n    - `\"db/1234567\"`\n    - `\"cod/2000000\"`\n   \
            \ - `\"cod/2000000@1234567\"`\n    - `\"nomad/L1234567890\"`\n    - `\"\
            42\"`"
          title: Id
          type: string
        links:
          allOf:
          - $ref: '#/components/schemas/ResourceLinks'
          description: a links object containing links related to the resource.
          title: Links
        meta:
          allOf:
          - $ref: '#/components/schemas/Meta'
          description: a meta object containing non-standard meta-information about
            a resource that can not be represented as an attribute or relationship.
          title: Meta
        relationships:
          allOf:
          - $ref: '#/components/schemas/EntryRelationships'
          description: 'A dictionary containing references to other entries according
            to the description in section Relationships encoded as [JSON API Relationships](https://jsonapi.org/format/1.0/#document-resource-object-relationships).

            The OPTIONAL human-readable description of the relationship MAY be provided
            in the `description` field inside the `meta` dictionary of the JSON API
            resource identifier object.'
          title: Relationships
        type:
          description: The name of the type of an entry.
          pattern: ^queries$
          title: Type
          type: string
      required:
      - id
      - type
      - attributes
      title: QueryResource
      type: object
...
  securitySchemes:
    OAuth2AuthorizationCodeBearer:
      flows:
        authorizationCode:
          authorizationUrl: https://staging.the-marketplace.eu/oauth/oauth2/auth
          scopes: {}
          tokenUrl: https://staging.the-marketplace.eu/oauth/oauth2/token
      type: oauth2
info:
  description: A gateway server to query multiple OPTIMADE databases.
  title: OPTIMADE Gateway
  version: 0.2.2
  x-application-id: ''
  x-application-name: OPTIMADE Gateway
  x-external-hostname: http://localhost:8000
  x-products:
  - name: OPTIMADE Gateway
    productId: 22e6e756-7179-47c9-8f37-1ee06c7b38ae
openapi: 3.0.2
paths:
...
  /queries:
    get:
      description: '`GET /queries`


        Return overview of all (active) queries.'
      operationId: get_queries_queries_get
      parameters:
      - description: A filter string, in the format described in section API Filtering
          Format Specification of the specification.
        in: query
        name: filter
        required: false
        schema:
          default: ''
          description: A filter string, in the format described in section API Filtering
            Format Specification of the specification.
          title: Filter
          type: string
      - description: 'The output format requested (see section Response Format).

          Defaults to the format string ''json'', which specifies the standard output
          format described in this specification.

          Example: `http://example.com/v1/structures?response_format=xml`'
        in: query
        name: response_format
        required: false
        schema:
          default: json
          description: 'The output format requested (see section Response Format).

            Defaults to the format string ''json'', which specifies the standard output
            format described in this specification.

            Example: `http://example.com/v1/structures?response_format=xml`'
          title: Response Format
          type: string
      - description: 'An email address of the user making the request.

          The email SHOULD be that of a person and not an automatic system.

          Example: `http://example.com/v1/structures?email_address=user@example.com`'
        in: query
        name: email_address
        required: false
        schema:
          default: ''
          description: 'An email address of the user making the request.

            The email SHOULD be that of a person and not an automatic system.

            Example: `http://example.com/v1/structures?email_address=user@example.com`'
          format: email
          title: Email Address
          type: string
      - description: 'A comma-delimited set of fields to be provided in the output.

          If provided, these fields MUST be returned along with the REQUIRED fields.

          Other OPTIONAL fields MUST NOT be returned when this parameter is present.

          Example: `http://example.com/v1/structures?response_fields=last_modified,nsites`'
        in: query
        name: response_fields
        required: false
        schema:
          default: ''
          description: 'A comma-delimited set of fields to be provided in the output.

            If provided, these fields MUST be returned along with the REQUIRED fields.

            Other OPTIONAL fields MUST NOT be returned when this parameter is present.

            Example: `http://example.com/v1/structures?response_fields=last_modified,nsites`'
          pattern: ([a-z_][a-z_0-9]*(,[a-z_][a-z_0-9]*)*)?
          title: Response Fields
          type: string
      - description: 'If supporting sortable queries, an implementation MUST use the
          `sort` query parameter with format as specified by [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-sorting).


          An implementation MAY support multiple sort fields for a single query.

          If it does, it again MUST conform to the JSON API 1.0 specification.


          If an implementation supports sorting for an entry listing endpoint, then
          the `/info/<entries>` endpoint MUST include, for each field name `<fieldname>`
          in its `data.properties.<fieldname>` response value that can be used for
          sorting, the key `sortable` with value `true`.

          If a field name under an entry listing endpoint supporting sorting cannot
          be used for sorting, the server MUST either leave out the `sortable` key
          or set it equal to `false` for the specific field name.

          The set of field names, with `sortable` equal to `true` are allowed to be
          used in the "sort fields" list according to its definition in the JSON API
          1.0 specification.

          The field `sortable` is in addition to each property description and other
          OPTIONAL fields.

          An example is shown in the section Entry Listing Info Endpoints.'
        in: query
        name: sort
        required: false
        schema:
          default: ''
          description: 'If supporting sortable queries, an implementation MUST use
            the `sort` query parameter with format as specified by [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-sorting).


            An implementation MAY support multiple sort fields for a single query.

            If it does, it again MUST conform to the JSON API 1.0 specification.


            If an implementation supports sorting for an entry listing endpoint, then
            the `/info/<entries>` endpoint MUST include, for each field name `<fieldname>`
            in its `data.properties.<fieldname>` response value that can be used for
            sorting, the key `sortable` with value `true`.

            If a field name under an entry listing endpoint supporting sorting cannot
            be used for sorting, the server MUST either leave out the `sortable` key
            or set it equal to `false` for the specific field name.

            The set of field names, with `sortable` equal to `true` are allowed to
            be used in the "sort fields" list according to its definition in the JSON
            API 1.0 specification.

            The field `sortable` is in addition to each property description and other
            OPTIONAL fields.

            An example is shown in the section Entry Listing Info Endpoints.'
          pattern: ([a-z_][a-z_0-9]*(,[a-z_][a-z_0-9]*)*)?
          title: Sort
          type: string
      - description: 'Sets a numerical limit on the number of entries returned.

          See [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-pagination).

          The API implementation MUST return no more than the number specified.

          It MAY return fewer.

          The database MAY have a maximum limit and not accept larger numbers (in
          which case an error code -- 403 Forbidden -- MUST be returned).

          The default limit value is up to the API implementation to decide.

          Example: `http://example.com/optimade/v1/structures?page_limit=100`'
        in: query
        name: page_limit
        required: false
        schema:
          default: 20
          description: 'Sets a numerical limit on the number of entries returned.

            See [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-pagination).

            The API implementation MUST return no more than the number specified.

            It MAY return fewer.

            The database MAY have a maximum limit and not accept larger numbers (in
            which case an error code -- 403 Forbidden -- MUST be returned).

            The default limit value is up to the API implementation to decide.

            Example: `http://example.com/optimade/v1/structures?page_limit=100`'
          minimum: 0.0
          title: Page Limit
          type: integer
      - description: 'RECOMMENDED for use with _offset-based_ pagination: using `page_offset`
          and `page_limit` is RECOMMENDED.

          Example: Skip 50 structures and fetch up to 100: `/structures?page_offset=50&page_limit=100`.'
        in: query
        name: page_offset
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _offset-based_ pagination: using
            `page_offset` and `page_limit` is RECOMMENDED.

            Example: Skip 50 structures and fetch up to 100: `/structures?page_offset=50&page_limit=100`.'
          minimum: 0.0
          title: Page Offset
          type: integer
      - description: 'RECOMMENDED for use with _page-based_ pagination: using `page_number`
          and `page_limit` is RECOMMENDED.

          It is RECOMMENDED that the first page has number 1, i.e., that `page_number`
          is 1-based.

          Example: Fetch page 2 of up to 50 structures per page: `/structures?page_number=2&page_limit=50`.'
        in: query
        name: page_number
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _page-based_ pagination: using `page_number`
            and `page_limit` is RECOMMENDED.

            It is RECOMMENDED that the first page has number 1, i.e., that `page_number`
            is 1-based.

            Example: Fetch page 2 of up to 50 structures per page: `/structures?page_number=2&page_limit=50`.'
          minimum: 0.0
          title: Page Number
          type: integer
      - description: 'RECOMMENDED for use with _cursor-based_ pagination: using `page_cursor`
          and `page_limit` is RECOMMENDED.'
        in: query
        name: page_cursor
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _cursor-based_ pagination: using
            `page_cursor` and `page_limit` is RECOMMENDED.'
          minimum: 0.0
          title: Page Cursor
          type: integer
      - description: 'RECOMMENDED for use with _value-based_ pagination: using `page_above`/`page_below`
          and `page_limit` is RECOMMENDED.

          Example: Fetch up to 100 structures above sort-field value 4000 (in this
          example, server chooses to fetch results sorted by increasing `id`, so `page_above`
          value refers to an `id` value): `/structures?page_above=4000&page_limit=100`.'
        in: query
        name: page_above
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _value-based_ pagination: using `page_above`/`page_below`
            and `page_limit` is RECOMMENDED.

            Example: Fetch up to 100 structures above sort-field value 4000 (in this
            example, server chooses to fetch results sorted by increasing `id`, so
            `page_above` value refers to an `id` value): `/structures?page_above=4000&page_limit=100`.'
          minimum: 0.0
          title: Page Above
          type: integer
      - description: 'RECOMMENDED for use with _value-based_ pagination: using `page_above`/`page_below`
          and `page_limit` is RECOMMENDED.'
        in: query
        name: page_below
        required: false
        schema:
          default: 0
          description: 'RECOMMENDED for use with _value-based_ pagination: using `page_above`/`page_below`
            and `page_limit` is RECOMMENDED.'
          minimum: 0.0
          title: Page Below
          type: integer
      - description: 'A server MAY implement the JSON API concept of returning [compound
          documents](https://jsonapi.org/format/1.0/#document-compound-documents)
          by utilizing the `include` query parameter as specified by [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-includes).


          All related resource objects MUST be returned as part of an array value
          for the top-level `included` field, see the section JSON Response Schema:
          Common Fields.


          The value of `include` MUST be a comma-separated list of "relationship paths",
          as defined in the [JSON API](https://jsonapi.org/format/1.0/#fetching-includes).

          If relationship paths are not supported, or a server is unable to identify
          a relationship path a `400 Bad Request` response MUST be made.


          The **default value** for `include` is `references`.

          This means `references` entries MUST always be included under the top-level
          field `included` as default, since a server assumes if `include` is not
          specified by a client in the request, it is still specified as `include=references`.

          Note, if a client explicitly specifies `include` and leaves out `references`,
          `references` resource objects MUST NOT be included under the top-level field
          `included`, as per the definition of `included`, see section JSON Response
          Schema: Common Fields.


          > **Note**: A query with the parameter `include` set to the empty string
          means no related resource objects are to be returned under the top-level
          field `included`.'
        in: query
        name: include
        required: false
        schema:
          default: references
          description: 'A server MAY implement the JSON API concept of returning [compound
            documents](https://jsonapi.org/format/1.0/#document-compound-documents)
            by utilizing the `include` query parameter as specified by [JSON API 1.0](https://jsonapi.org/format/1.0/#fetching-includes).


            All related resource objects MUST be returned as part of an array value
            for the top-level `included` field, see the section JSON Response Schema:
            Common Fields.


            The value of `include` MUST be a comma-separated list of "relationship
            paths", as defined in the [JSON API](https://jsonapi.org/format/1.0/#fetching-includes).

            If relationship paths are not supported, or a server is unable to identify
            a relationship path a `400 Bad Request` response MUST be made.


            The **default value** for `include` is `references`.

            This means `references` entries MUST always be included under the top-level
            field `included` as default, since a server assumes if `include` is not
            specified by a client in the request, it is still specified as `include=references`.

            Note, if a client explicitly specifies `include` and leaves out `references`,
            `references` resource objects MUST NOT be included under the top-level
            field `included`, as per the definition of `included`, see section JSON
            Response Schema: Common Fields.


            > **Note**: A query with the parameter `include` set to the empty string
            means no related resource objects are to be returned under the top-level
            field `included`.'
          title: Include
          type: string
      - description: If the client provides the parameter, the value SHOULD have the
          format `vMAJOR` or `vMAJOR.MINOR`, where MAJOR is a major version and MINOR
          is a minor version of the API. For example, if a client appends `api_hint=v1.0`
          to the query string, the hint provided is for major version 1 and minor
          version 0.
        in: query
        name: api_hint
        required: false
        schema:
          default: ''
          description: If the client provides the parameter, the value SHOULD have
            the format `vMAJOR` or `vMAJOR.MINOR`, where MAJOR is a major version
            and MINOR is a minor version of the API. For example, if a client appends
            `api_hint=v1.0` to the query string, the hint provided is for major version
            1 and minor version 0.
          pattern: (v[0-9]+(\.[0-9]+)?)?
          title: Api Hint
          type: string
      responses:
        '200':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/QueriesResponse'
          description: Successful Response
        '400':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '553':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Version Not Supported
      security:
      - OAuth2AuthorizationCodeBearer: []
      summary: Get Queries
      tags:
      - Queries
    post:
      description: '`POST /queries`


        Create or return existing gateway query according to `query`.'
      operationId: post_queries_queries_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QueryCreate'
        required: true
      responses:
        '202':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/QueriesResponseSingle'
          description: Successful Response
        '400':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '553':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Version Not Supported
      security:
      - OAuth2AuthorizationCodeBearer: []
      summary: Post Queries
      tags:
      - Queries
  /queries/{query_id}:
    get:
      description: '`GET /queries/{query_id}`


        Return a single [`QueryResource`][optimade_gateway.models.queries.QueryResource].'
      operationId: get_query_queries__query_id__get
      parameters:
      - in: path
        name: query_id
        required: true
        schema:
          title: Query Id
          type: string
      responses:
        '200':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/QueriesResponseSingle'
          description: Successful Response
        '400':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Bad Request
        '401':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unauthorized
        '403':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Forbidden
        '404':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Found
        '422':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Unprocessable Entity
        '500':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Internal Server Error
        '501':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Not Implemented
        '553':
          content:
            application/vnd.api+json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
          description: Version Not Supported
      security:
      - OAuth2AuthorizationCodeBearer: []
      summary: Get Query
      tags:
      - Queries
...
servers:
- url: http://localhost:8000